% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSAT.R
\docType{package}
\name{CSAT-package}
\alias{CSAT}
\alias{CSAT-package}
\title{CSAT: California Stream Assessment Toolbox}
\description{
Combined toolbox for CSCI, ASCI, and PHAB.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/SCCWRP/CSAT/issues}
}

}
\author{
\strong{Maintainer}: Marcus W. Beck \email{marcusb@sccwrp.org}

}
\keyword{internal}
