% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{CSAT_conflicts}
\alias{CSAT_conflicts}
\title{Conflicts between CSAT and other packages}
\usage{
CSAT_conflicts()
}
\description{
This function lists all the conflicts between packages in CSAT
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
CSAT_conflicts()
}
