% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{CSAT_deps}
\alias{CSAT_deps}
\title{List all CSAT dependencies}
\usage{
CSAT_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
CSAT packages.}
}
\description{
List all CSAT dependencies
}
