% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CSAT_packages}
\alias{CSAT_packages}
\title{List all packages in CSAT}
\usage{
CSAT_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include CSAT in the list?}
}
\description{
List all packages in CSAT
}
\examples{
CSAT_packages()
}
