% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{CSAT_update}
\alias{CSAT_update}
\title{Update CSAT packages}
\usage{
CSAT_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
CSAT packages.}
}
\description{
This will check to see if all CSAT packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
CSAT_update()
}
}
